Hack To Detect End Of Assault Games (#25)
I hate everything about this hack, but i think its the best way to fix
Assault without introducing more complexity.

One alternative i thought about was introducing an interface package
IVoteSys with a GameProxy class that has a static function IsGameOver
that could be used after dynamically loading a derived class suitable
for a given GameInfo subclass.
Problem was how to specify the Proxy class in the server config and the
general complexity of this type of infrastructure.

Another alternative was having the proxy class be instantiated by
VoteSys and then having it monitor Level.Game for the actual end. The
proxy would then set bGameEnded to True for itself. VoteSys would then
use GetPropertyText to query the proxy.
This reduces complexity for third parties (no interface package, no
deriving from specific class, etc.), but there is still to my knowledge
only one use-case for it: Assault.
This also doesnt address the configuration complexity for server admins.

So here is what was necessary to make Assault work. Perhaps removing
MultiRoundGameModes would be best in order to not confuse people.