Fix Windows Not Giving Back Mouse Control
Some mods replace the games console with their own (UTPure, NewNet,
UltimateNewNet, ...). The replacement consoles dont make use of states,
and instead have their own variable ("zzMyState") to keep track of the
state.

This commit introduces a check of the variable in addition to the state
name when deciding whether to automatically close menuing.